+++ docs/changelog.rst
@@ -3,8 +3,8 @@
 Changelog
 =========
 
-v1.6.0 (in development)
------------------------
+v1.6.0 (2024-07-18)
+-------------------
 - Drop support for Python 3.7
 
 - Support :pep:`740`